converter/requirements.txt
absl-py==1.3.0
appnope==0.1.3
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
cachetools==5.2.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.1.1
contourpy==1.0.6
coremltools==6.1
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
flatbuffers==22.11.23
fonttools==4.38.0
gast==0.4.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.7.0
idna==3.4
importlib-metadata==5.0.0
ipython==8.10.0
jedi==0.18.2
keras==2.11.0
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mpmath==1.3.0
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.1
pycparser==2.21
Pygments==2.15.0
pyparsing==3.0.9
pytest==7.4.0
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
seaborn==0.12.1
six==1.16.0
sounddevice==0.4.5
stack-data==0.6.2
sympy==1.11.1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.1
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
tflite-support-nightly==0.4.4.dev20221103
torch==1.13.1
torchvision==0.14.0
tqdm==4.64.1
traitlets==5.8.1
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.10.0
